+++ Mods ChoGGi/Reorder Queued Research/metadata.lua
@@ -1,12 +1,20 @@
 return PlaceObj("ModDef", {
+	"dependencies", {
+		PlaceObj("ModDependency", {
+			"id", "ChoGGi_Library",
+			"title", "ChoGGi's Library",
+			"version_major", 9,
+			"version_minor", 6,
+		}),
+	},
 	"title", "Reorder Queued Research",
 	"id", "ChoGGi_ReorderQueuedResearch",
 	"steam_id", "2449037850",
 	"pops_any_uuid", "8ec18d2c-3035-4f7a-8af3-3a6f14507bc0",
 	"lua_revision", 1001569,
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -16,10 +24,5 @@ return PlaceObj("ModDef", {
 	"TagInterface", true,
 	"description", [[
 Reorder the research queue with crappy looking buttons.
-
-
-Known Issues:
-Didn't bother finding out a way to update the list, so it closes and reopens the dialog.
-It'll take a couple milliseconds, so you'll see a blink and the background animation gets reset.
 ]],
 })